refactor: App.js에 유저 정보만 불러오기 및 라우터 컴포넌트 재설정
