--- src/css/App.css
@@ -1,8 +1,0 @@
-/* main */
-main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 2rem;
-  gap: 3.2rem;
-}
